refactor(schematic): split schematic between code-runtime & code-schematic
