+++ code/code-schematics/src/schematic/rules/index.ts
