+++ code/code-schematics/src/schematic/utils/gitignore.utils.ts
